+++ tasks/seq/korobeinikov_matrix_multiplication_horizontal_scheme_A_vertical_scheme_B/func_tests/main_korobeinikov.cpp
@@ -48,7 +48,7 @@ TEST(matrix_multiplication_horizontal_scheme_A_vertical_scheme_B, Test_1_without
   testTaskSequential.pre_processing();
   testTaskSequential.run();
   testTaskSequential.post_processing();
-  for (int i = 0; i < right_answer.size(); i++) {
+  for (size_t i = 0; i < right_answer.size(); i++) {
     ASSERT_EQ(right_answer[i], out[i]);
   }
   ASSERT_EQ(count_rows_out, count_rows_RA);
@@ -100,7 +100,7 @@ TEST(matrix_multiplication_horizontal_scheme_A_vertical_scheme_B, Test_2_with_ne
   testTaskSequential.pre_processing();
   testTaskSequential.run();
   testTaskSequential.post_processing();
-  for (int i = 0; i < right_answer.size(); i++) {
+  for (size_t i = 0; i < right_answer.size(); i++) {
     ASSERT_EQ(right_answer[i], out[i]);
   }
   ASSERT_EQ(count_rows_out, count_rows_RA);
@@ -152,7 +152,7 @@ TEST(matrix_multiplication_horizontal_scheme_A_vertical_scheme_B, Test_3_only_ze
   testTaskSequential.pre_processing();
   testTaskSequential.run();
   testTaskSequential.post_processing();
-  for (int i = 0; i < right_answer.size(); i++) {
+  for (size_t i = 0; i < right_answer.size(); i++) {
     ASSERT_EQ(right_answer[i], out[i]);
   }
   ASSERT_EQ(count_rows_out, count_rows_RA);
